Finish implementing all needed calls to Metro ArcGIS API
This simplifies our integration with Metro, and removes the autocomplete endpoint. This is also a WIP that removes all our useful endpoints. The architecture is now:

/landuses/?address="" - an endpoint that returns _all_ land uses for an address, and the zone of the address. or 422 if the address isn't a valid Nashville address.
/landusecategories - a list of string land use categories.

These have to be added back in.  The `address_candidates` has been replaced by a `geocode_address` client that just returns the first geocoded point for a candidate address string.  In addition, instead of looking up the ParcelID and then using that to get zoning history, this looks up the zone directly using the geocoded point.

Tests have been added for the internal api integration - they are only run if requested explicitly and hit the actual metro api. A note has been added to the README - they can be run using `mix test --include external`.