Finish ingest: Write data to Postgres using Ecto
This finishes the ingestion script, modifying it to write data to Postgres using Ecto.

This makes a host of other improvements and while we need to QA the output more, the ingestion should be basically working. I'm not using Ecto well (I skimmed the docs when I should have just sat down and read them) but it works.