Finish ingest: Write data to Postgres using Ecto (#49)
This finishes the ingestion script, modifying it to write data to Postgres using Ecto.

This makes a host of other improvements and while we need to QA the output more, the ingestion should be basically working. I'm not using Ecto well (I skimmed the docs when I should have just sat down and read them) but it works.

It also updates he README and Travis with instructions on settings up a PostgreSQL 9.6 database. Going with 9.6 for now because Travis is unable to run 10 w/out an extra script (travis-ci/travis-ci#8537).

This also simplifies the internal naming. The onerous ZoningDistrict and ZoningDistrictLandUseCondition have been renamed to Zone and ZoneLandUseCondition respectively.